[arm,pydgin] some work on refactoring arm-sim.py
works on cpython interp, nojit translation
